feat (S11P31D102-232) : distance Integer -> Double로 변경

+++ backend/runIt/src/main/java/com/ssafy/runit/domain/record/dto/response/RecordTodayResponse.java
@@ -4,11 +4,11 @@ import lombok.Builder;
 
 @Builder
 public record RecordTodayResponse(
-        Integer distance,
+        Double distance,
         Long time,
         Integer pace
 ) {
-    public static RecordTodayResponse fromEntity(Integer dis, Long time, Integer pace){
+    public static RecordTodayResponse fromEntity(Double dis, Long time, Integer pace){
         return RecordTodayResponse.builder()
                 .pace(pace)
                 .time(time)
